fix: client side hook errors
